+++ targets/nuttx-stm32f4/Makefile
@@ -33,10 +33,11 @@ CFLAGS += -I$(JERRYSCRIPT_ROOT_DIR)/jerry-ext/include
 CFLAGS += -I$(JERRYSCRIPT_ROOT_DIR)/jerry-libm/include
 
 # These libs should be copied from the JerryScript project.
-LIBS = libjerry-core.a libjerry-ext.a libjerry-port-default.a libjerry-port-default-minimal.a libjerry-libm.a
+LIBS = libjerry-core.a libjerry-ext.a libjerry-libm.a
 
 APPNAME = jerry
 ASRCS = setjmp.S
+CSRCS = jerry_port.c
 MAINSRC = jerry_main.c
 
 .PHONY: copylibs
