+++ targets/nuttx-stm32f4/Makefile.travis
@@ -51,7 +51,7 @@ install: install-apt-get-deps install-noapt
 
 # Build JerryScript.
 script-build-jerryscript:
-	tools/build.py --clean --toolchain cmake/toolchain_mcu_stm32f4.cmake --profile=es2015-subset --jerry-cmdline OFF --lto OFF --jerry-libm ON --all-in-one ON --mem-heap 70 --compile-flag='--sysroot=../nuttx'
+	tools/build.py --clean --toolchain cmake/toolchain_mcu_stm32f4.cmake --profile=es2015-subset --jerry-cmdline OFF --lto OFF --jerry-libm ON --all-in-one ON --jerry-port-default OFF --mem-heap 70 --compile-flag='--sysroot=../nuttx'
 
 # Link in the NuttX JerryScript target directory under the NuttX apps tree.
 script-add-jerryscript-app:
